+++ pyjwt_key_fetcher/provider.py
@@ -2,31 +2,50 @@ from typing import Any, Dict, Optional
 
 import aiocache  # type: ignore
 
-from pyjwt_key_fetcher.errors import JWTKeyNotFoundError, JWTOpenIDConnectError
+from pyjwt_key_fetcher.errors import (
+    JWTKeyNotFoundError,
+    JWTProviderConfigError,
+    JWTProviderJWKSError,
+)
 from pyjwt_key_fetcher.http_client import HTTPClient
 from pyjwt_key_fetcher.key import Key
 
 
-class OpenIDProvider:
-    def __init__(self, iss: str, http_client: HTTPClient) -> None:
+class Provider:
+    def __init__(
+        self,
+        iss: str,
+        http_client: HTTPClient,
+        config_path: str = "/.well-known/openid-configuration",
+    ) -> None:
         self.iss = iss
         self.http_client = http_client
-        self._openid_configuration: Optional[Dict[str, Any]] = None
+        self._configuration: Optional[Dict[str, Any]] = None
         self._jwk_map: Dict[str, Dict[str, Any]] = {}
         self.keys: Dict[str, Key] = {}
+        self.config_path = config_path
 
-    async def get_openid_configuration(self) -> Dict[str, Any]:
+    async def _config_uri(self) -> str:
         """
-        Get the OpenID configuration.
+        Get the URI at which the configuration is expected to be found.
 
-        :return: The OpenID Configuration as a dictionary.
+        Can be for example https://example.com/.well-known/openid-configuration
+        :return: The configuration URI.
+        """
+        return f"{self.iss.rstrip('/')}{self.config_path}"
+
+    async def get_configuration(self) -> Dict[str, Any]:
+        """
+        Get the configuration as a dictionary.
+
+        :return: The configuration as a dictionary.
         :raise JWTHTTPFetchError: If there's a problem fetching the data.
         """
-        if self._openid_configuration is None:
-            url = f"{self.iss.rstrip('/')}/.well-known/openid-configuration"
-            self._openid_configuration = await self.http_client.get_json(url)
+        if self._configuration is None:
+            url = await self._config_uri()
+            self._configuration = await self.http_client.get_json(url)
 
-        return self._openid_configuration
+        return self._configuration
 
     async def _get_jwks_uri(self) -> str:
         """
@@ -34,15 +53,13 @@ class OpenIDProvider:
 
         :return: The uri to the JWKs.
         :raise JWTHTTPFetchError: If there's a problem fetching the data.
-        :raise JWTOpenIDConnectError: If the data doesn't contain "jwks_uri".
+        :raise JWTProviderConfigError: If the config doesn't contain "jwks_uri".
         """
-        conf = await self.get_openid_configuration()
+        conf = await self.get_configuration()
         try:
             jwks_uri = conf["jwks_uri"]
         except KeyError as e:
-            raise JWTOpenIDConnectError(
-                "Missing 'jwks_uri' in OpenID Connect configuration"
-            ) from e
+            raise JWTProviderConfigError("Missing 'jwks_uri' in configuration") from e
         return jwks_uri
 
     @aiocache.cached(ttl=300)
@@ -54,14 +71,15 @@ class OpenIDProvider:
 
         :return: A mapping of {kid: {<data_for_the_kid>}, ...}
         :raise JWTHTTPFetchError: If there's a problem fetching the data.
-        :raise JWTOpenIDConnectError: If the data doesn't contain "jwks_uri".
+        :raise JWTProviderConfigError: If the config doesn't contain "jwks_uri".
+        :raise JWTProviderJWKSError: If the jwks_uri is missing the "jwks".
         """
         jwks_uri = await self._get_jwks_uri()
         data = await self.http_client.get_json(jwks_uri)
         try:
             jwks_list = data["keys"]
         except KeyError as e:
-            raise JWTOpenIDConnectError(f"Missing 'jwks' in {jwks_uri}") from e
+            raise JWTProviderJWKSError(f"Missing 'jwks' in {jwks_uri}") from e
 
         jwk_map = {jwk["kid"]: jwk for jwk in jwks_list}
 
@@ -74,7 +92,8 @@ class OpenIDProvider:
         :param kid: The key ID.
         :return: The raw JWK data as a dictionary.
         :raise JWTHTTPFetchError: If there's a problem fetching the data.
-        :raise JWTOpenIDConnectError: If the data doesn't contain "jwks_uri".
+        :raise JWTProviderConfigError: If the config doesn't contain "jwks_uri".
+        :raise JWTProviderJWKSError: If the jwks_uri is missing the "jwks".
         :raise JWTKeyNotFoundError: If no matching kid was found.
         """
         if kid not in self._jwk_map:
@@ -91,7 +110,8 @@ class OpenIDProvider:
         :param kid: The key id.
         :return: The Key.
         :raise JWTHTTPFetchError: If there's a problem fetching the data.
-        :raise JWTOpenIDConnectError: If the data doesn't contain "jwks_uri".
+        :raise JWTProviderConfigError: If the config doesn't contain "jwks_uri".
+        :raise JWTProviderJWKSError: If the jwks_uri is missing the "jwks".
         :raise JWTKeyNotFoundError: If no matching kid was found.
         """
         if kid not in self.keys: